+++ framework/domain/src/test/java/is/codion/framework/domain/entity/DefaultEntityTest.java
@@ -475,10 +475,9 @@ public class DefaultEntityTest {
     assertTrue(testEntity.isNull(Detail.MASTER_FK));
     testEntity.put(Detail.MASTER_ID, 10L);
 
-    assertFalse(testEntity.loaded(Detail.MASTER_FK));
+    assertNull(testEntity.get(Detail.MASTER_FK));
     Entity referencedEntityValue = testEntity.referencedEntity(Detail.MASTER_FK);
     assertEquals(10L, referencedEntityValue.get(Master.ID));
-    assertFalse(testEntity.loaded(Detail.MASTER_FK));
     assertFalse(testEntity.isNull(Detail.MASTER_FK));
     assertFalse(testEntity.isNull(Detail.MASTER_ID));
 
@@ -732,7 +731,6 @@ public class DefaultEntityTest {
     assertEquals(1, emp.get(Employee.DEPARTMENT_NO));
     emp.put(Employee.DEPARTMENT_NO, 2);
     assertNull(emp.get(Employee.DEPARTMENT_FK));
-    assertFalse(emp.loaded(Employee.DEPARTMENT_FK));
     Entity referencedDept = emp.referencedEntity(Employee.DEPARTMENT_FK);
     assertEquals(Integer.valueOf(2), referencedDept.primaryKey().get());
 
@@ -768,10 +766,10 @@ public class DefaultEntityTest {
             .valuesEqual(copy.referencedEntity(Employee.MANAGER_FK).referencedEntity(Employee.DEPARTMENT_FK)));
 
     emp.save();
-    emp.definition().foreignKeys().get().forEach(foreignKey -> assertTrue(emp.loaded(foreignKey)));
+    emp.definition().foreignKeys().get().forEach(foreignKey -> assertNotNull(emp.get(foreignKey)));
     emp.definition().foreignKeys().get().forEach(emp::remove);
     assertFalse(emp.modified());
-    emp.definition().foreignKeys().get().forEach(foreignKey -> assertFalse(emp.loaded(foreignKey)));
+    emp.definition().foreignKeys().get().forEach(foreignKey -> assertNull(emp.get(foreignKey)));
   }
 
   @Test