+++ framework/model/src/test/java/is/codion/framework/model/AbstractEntityEditModelTest.java
@@ -181,7 +181,7 @@ public final class AbstractEntityEditModelTest {
     assertTrue(employeeEditModel.isNull(Employee.DEPARTMENT_FK).get());
     assertFalse(employeeEditModel.isNotNull(Employee.DEPARTMENT_FK).get());
     employeeEditModel.put(Employee.DEPARTMENT, dept.get(Department.ID));
-    assertFalse(employeeEditModel.entity().loaded(Employee.DEPARTMENT_FK));
+    assertNull(employeeEditModel.entity().get(Employee.DEPARTMENT_FK));
     dept = employeeEditModel.get(Employee.DEPARTMENT_FK);
     assertNull(dept);
     employeeEditModel.setDefaults();